Update Auto-DevOps-template Environment Variables
Added two new ENV vars to template:
 - HELM_EXTERNAL_PORT map to helm parameter service.externalPort
 - HELM_EXTERNAL_PORT map to helm parameter service.internalPort

This will allow users to set the TCP port that the container listens on internally and externally when deploying to Kubernetes.

Currently the static default is 5000 and can only be changed by adding additional `--set service.externalPort` and `--set service.internalPort` to AutoDevOps .gitlab-ci.yml file.