Optimize MergeRequestDiff#last_commit_sha
The previous implementation would load all the MergeRequestDiffCommit
entries into memory and retrieve the SHA from the latest entry. We can
save a bit of query time and memory by retrieving the SHA directly from
the first item from the database. This is particularly helpful when
there are many merge requests that need this information.

Seen while investigating Todo API performance in
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/25711.