Fix failing spec in spec/features/users/login_spec.rb
This spec would start to fail on the first of the month because the
generated text would have a zero-padded day (e.g. 01 vs 1), whereas the
expected text check used an unpadded day via the `-d` parameter in
strftime (https://apidock.com/ruby/Date/strftime). To fix this, we
use use the `d` parameter to pad zeros.

This spec was introduced recently in
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/25731.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/59731