Route to emptyState straight from router
As soon as we detect that the project is empty, we can route to empty
state immediately without the attempt to load a branch even if it's
specified in the URL.

This means that is we got to fetching a branch and it fails, we're in
non-empty repo and the branch really doesn't exist. Hence we should show
the error.