Fix project settings not giving updates upon failure
When project settings cannot be updated, the AJAX request would returned
a snippet of JavaScript that was never displayed. To fix this, we force
the controller to redirect to the edit page by disabling remote fields.
It turns out in this section we don't use remote fields in any case.