Resolves race condition for mergeability check service
The MergeabilityCheckService can return success even with
an outdated merge ref. That's because the
MergeRequests::RefreshService is called async and it has
the responsibility of flipping the MRs merge_status to
unchecked when they receive updates, or its target branches
receives an update. Now, for callers that need instant
feedback from the repository, we pass a recheck flag to
the service, which will both sync merge-status with
merge-ref altogether, if the ref is outdated.

This can't be done for every scenario given it's an
expensive calculation.