Run tests in parallel via parallel_tests
Uses the parallel_tests gem to execute tests in multiple processes
simultaneously on the same machine.

Adds the `--parallel` CLI option that instructs the QA framework
to use the parallel_tests executable.

Tests need access to global state contained in `Runtime::Scenario`
so when `--parallel` is invoked `Runtime::Scenario` is serialized
to an environment variable, which is passed to parallel_tests,
and then deserialized in `spec_helper`.

Configure Browser/Capybara in spec_helper

This is one requirement for running tests via the parallel_tests gem,
which also contributes to running the E2E test directly via rspec
(instead of bin/qa)

Without this change the configuration is only loaded when bin/qa is
run, but not when Specs::Runner spawns new processes to actually
execute the tests.

Serialize parallel_tests output

This makes it possible to read the test logs.

Fix failing unit tests

Honor parallel_tests exit status

This means the QA framework will exit with an appropriate exit status
if any of the tests fail

Pass environment variables to parallel_tests

Create personal access token for parallel_tests

If no token is passed as an environment variable, each parallel
process would create its own token. This change creates one
token and passes it to each process as an environment variable.

Import ENV_VARIABLES from gitlab-qa

Refactor Specs::Runner to reduce complexity

Extract a couple of methods for knapsack- and rspec-related code.
Extract Specs::ParallelRunner to its own module.

Apply code review suggestions

Update gitlab-qa to 4.0.0

This version includes the --parallel option and LOAD_PATH fixes

Run review-app tests via parallel_tests

Retry creating sandbox on failure

This is a workaround for a race condition where two processes can't
get a sandbox group so they both try to create one, and one fails
because the other finished first. With this change the process that
fails to create a sandbox group tries to get it again.