Speed up MR widget by disabling BatchLoader replace_methods
We've seen a significant performance penalty when using
`BatchLoader#__replace_with!`. This defines methods on the batch loader
that proxy to the 'real' object using send. The alternative is
`method_missing`, which is slower.  However, we've noticed that
`method_missing` can be faster if:

1. The objects being loaded have a large interface.
2. We don't call too many methods on the loaded object.

In proudction, we've observed a small percentage of time used
in Module#define_methods as a result of the BatchLoader. Disabling this
for ActiveRecord models is generally a good thing to do since there are
so many methods.