Add last_version, page_count and group_by_directory to WikiDirectory
This adds methods to support displaying number of child pages within a
directory, and the latest modified version of any page within that
directory.

This is a better place for group_by_directory since now `WikiDirectory`
refers to `WikiPage`, but `WikiPage` does not refer to `WikiDirectory`,
eliminating a circular topological sort. This also eliminates smells in
the specs by avoiding the need to refer to classes by constant names.