Disable setting of Sidekiq JID in Redis for some workers
In https://gitlab.com/gitlab-com/gl-infra/production/issues/939, we see
that the highest usage of Redis commands are from setting and deleting
the Sidekiq job ID from gitlab-sidekiq-status.

For most jobs, we don't need to store the job ID into Redis before
and after the job runs. We disable this on the most common workers to
reduce the load on Redis.