Migrations guide: use atomic steps, when possible
Currently, the DB migrations guide says that "you must" use non-blocking
operations (such as CREATE INDEX CONCURRENTLY), always. But this does not make
sense in cases of empty tables and leads to splitting the work to multiple
non-atomic (with disable_ddl_transaction!) DB migrations.

To follow KISS principle, to have fewer DB migrations steps, to have them
atomic when it's possible and simplify deployment and troubleshooting,
the following exceptions were added to the doc:

- index creation,
- index dropping,
- defining an FK,
- adding a column with DEFAULT,

saying that if the target table is empty, it's completely normal and
recommended to use regular, blocking methods (we will not block anyone if the
target table is empty).

Additionally, some grammar fixes and style improvements were added.