Assert SSH key was added before using it
The test would sometimes fail because it added an SSH key via the UI
but tried to use it before the server had completed adding it.

This also moves the code to delete the SSH key to an after hook so
that it runs even if the test fails midway, e.g., if it fails during
the push.

It also adds an `element_text` method to wrap Capybara's `.text` in
a method that we won't have to change later if we switch to a
different framework.

Finally, it adds an Assertions module that contains assertions we
can use where `expect` is not appropriate (and wraps Capybara
methods)