Only fetch non-default branch feedback
We only want to update feedback for non-default branch feedback with
this migration. (Default branch feedback will be updated in an
earlier migration.)

Directly checking if the related pipeline is on the default branch
is too difficult without using Gitaly related classes, and we try to
avoid using Gitlab classes that might change in migrations.

Therefore, I'm using the existence of occurrences related to the
pipeline as an indicator of the pipeline having run for the default
branch. (At the moment, we only create occurrences for pipelines on
the default branch.)