Fix tests that depend on wiki routes
This ensures that the test that depend on specific routes pass, that we
still support legacy routes, and that we obey both rails conventions and
GitLab rules concerning wildcards.

Address route precedence issues
The redirection needs to happen after routing to new. A test is added to
catch regressions.

This also makes use of `File.join` the default.