Fix project exports clobbering concurrent export paths
When a project export completes, it removes everything in
`Project#import_export_shared.archive_path`, which can erase files
needed for another ongoing project export. This is problematic for
custom templates, which exports an existing project to get the most
recent changes and imports that archive to another project.

To avoid this from happening, we generate a random unique subpath in the
shared temporary directory so that multiple exports can work at the same
time.

Closes https://gitlab.com/gitlab-org/gitlab-ee/issues/14716